location for dice

+++ miro/die/main.js
@@ -35,44 +35,55 @@ miro.onReady(() => {
           svgIcon: icon, //'<circle cx="12" cy="12" r="9" fill="none" fill-rule="evenodd" stroke="currentColor" stroke-width="2"/>',
           onClick: async () => {
 
-            var num = get_randome();
+            let location = await miro.board.selection.get()
 
-            let dice = (await miro.board.widgets.create(
-              { type:'sticker', 
-                text: num.toString(),
-                x: 0.0,
-                y: 0.0,
-                capabilities: {
-                "editable": false
-                },
-            }))[0];
-
-            for (let i = 0; i < 15; i++) 
+            if (location.length == 1)
             {
               var num = get_randome();
-              var color = get_random_color();
 
+              let dice = (await miro.board.widgets.create(
+                { type:'sticker', 
+                  text: num.toString(),
+                  x: location[0].x,
+                  y: location[0].y,
+                  capabilities: {
+                  "editable": false
+                  },
+              }))[0];
+  
+              for (let i = 0; i < 15; i++) 
+              {
+                var num = get_randome();
+                var color = get_random_color();
+  
+                await miro.board.widgets.update({
+                    id: dice.id, 
+                    text: num.toString(), 
+                    style:{
+                      stickerBackgroundColor: color,
+                    }
+                  }) // update sticker
+  
+                  sleep(50)
+              }
+  
               await miro.board.widgets.update({
-                  id: dice.id, 
-                  text: num.toString(), 
-                  style:{
-                    stickerBackgroundColor: color,
-                  }
-                }) // update sticker
-
-                sleep(50)
+                id: dice.id, 
+                text: num.toString(), 
+                style:{
+                  stickerBackgroundColor: '#5ee335',
+                }
+              }) // update sticker
+              
+              await sleep(5000);
+              await miro.board.widgets.deleteById(dice.id) // delete sticker
+            }
+            else
+            {
+              miro.showNotification('Please select ONE widget as location.')
             }
 
-            await miro.board.widgets.update({
-              id: dice.id, 
-              text: num.toString(), 
-              style:{
-                stickerBackgroundColor: '#5ee335',
-              }
-            }) // update sticker
-            
-            await sleep(5000);
-            await miro.board.widgets.deleteById(dice.id) // delete sticker
+           
           }
         }
       }
